rename some files and fix imports

+++ lib/model/cache_manager.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:desafio_tokenlab/models/api_service.dart';
+import 'package:desafio_tokenlab/model/api_service.dart';
 import 'package:path_provider/path_provider.dart';
 
 class CacheManager {
@@ -18,10 +18,10 @@ class CacheManager {
     return File('$localPath/prefs.json');
   }
 
-  Future<File> _writePrefsFile({required bool cached}) async {
+  void _writePrefsFile({required bool cached}) async {
     final prefsFile = await _prefsFile;
 
-    return prefsFile.writeAsString(json.encode({'cached': cached}));
+    await prefsFile.writeAsString(json.encode({'cached': cached}));
   }
 
   Future<File> get _moviesFile async {
@@ -29,9 +29,9 @@ class CacheManager {
     return File('$localPath/movies.json');
   }
 
-  Future<File> _writeMoviesFile(String data) async {
+  void _writeMoviesFile(String data) async {
     final file = await _moviesFile;
-    return file.writeAsString(data);
+    await file.writeAsString(data);
   }
 
   Future<bool> get isDataCached async {
@@ -45,9 +45,8 @@ class CacheManager {
     return prefsDataDecoded['cached'];
   }
 
-  Future<void> writeCache(String data) async {
+  void writeCache(String data) async {
     final List dataJson = json.decode(data);
-    // final List<Map<String,dynamic>> completeMovieData = [];
 
     for (var movie in dataJson) {
       final movieData = await ApiService.movieData(movie['id']);
@@ -55,21 +54,21 @@ class CacheManager {
       movie['description'] = movieDataDecoded['overview'];
     }
 
-    Future.wait<File>([
-      _writePrefsFile(cached: true),
-      _writeMoviesFile(json.encode(dataJson)),
-    ]);
+    _writePrefsFile(cached: true);
+    _writeMoviesFile(json.encode(dataJson));
   }
 
-  Future<void> get eraseCache async {
+  void get eraseCache async {
     final file = await _moviesFile;
+    final fileExists = await file.exists();
+    if (!fileExists) return;
     final cache = await data;
     final isCached = await isDataCached;
     final List cacheDecoded = json.decode(cache);
 
     if (isCached) {
       await file.delete();
-      await _writePrefsFile(cached: false);
+      _writePrefsFile(cached: false);
 
       for (var movie in cacheDecoded) {
         CachedNetworkImage.evictFromCache(movie['poster_url'] as String);
